% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_buettner.R
\docType{data}
\name{buettner}
\alias{buettner}
\alias{data_example}
\title{Real data example for CESME}
\format{
\describe{
 \item{Rows}{Objects}
 \item{Columns}{Genes}
}
}
\usage{
data(buettner)
}
\description{
The data contains gene expression data of 182 objects on 8989 genes.
}
\details{
This is the gene expression data from Buettner
}
\examples{
data(buettner)
require(kernlab)
require(bigmemory)
buettner_label = data_buettner[,1]
buettner_covariate = data_buettner[,-1]
n = dim(buettner_covariate)[1]
p = dim(buettner_covariate)[2]
K = max(buettner_label)
z.initial = as.numeric(specc(buettner_covariate, centers = K))
npn.def = npn.clust.bic.large(buettner_covariate, K=K)
npn.clust = npn.def$z[npn.def$iter+1,,1]
}
\keyword{dataset}
